# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Role"
							}
						],
						"nullable": true
					},
					"files": {
						"items": {
							"$ref": "#/components/schemas/File"
						},
						"type": "array"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updateAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"email",
					"name",
					"image",
					"password",
					"role",
					"files",
					"createAt",
					"updateAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"root": {
						"type": "boolean"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/Permissions"
						},
						"type": "array"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updateAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"root",
					"users",
					"permissions",
					"createAt",
					"updateAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Permissions": {
				"properties": {
					"id": {
						"type": "string"
					},
					"role_id": {
						"type": "string"
					},
					"resource": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"roles": {
						"$ref": "#/components/schemas/Role"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updateAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"role_id",
					"resource",
					"value",
					"roles",
					"createAt",
					"updateAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"File": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"disk": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"size": {
						"type": "string"
					},
					"checksum": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updateAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"disk",
					"name",
					"path",
					"type",
					"size",
					"checksum",
					"user",
					"createAt",
					"updateAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_User_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"errors": {},
					"data": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"code",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_User.Exclude_keyofUser.id-or-createAt-or-updateAt__": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"files": {
						"items": {
							"$ref": "#/components/schemas/File"
						},
						"type": "array"
					}
				},
				"required": [
					"username",
					"email",
					"name",
					"image",
					"password",
					"role",
					"files"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.id-or-createAt-or-updateAt_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.id-or-createAt-or-updateAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IUserPayload": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_User.id-or-createAt-or-updateAt_"
					},
					{
						"properties": {
							"roleId": {
								"type": "string"
							}
						},
						"required": [
							"roleId"
						],
						"type": "object"
					}
				]
			},
			"Pagination_User_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"per_page": {
						"type": "number",
						"format": "double"
					},
					"current_page": {
						"type": "number",
						"format": "double"
					},
					"last_page": {
						"type": "number",
						"format": "double"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"total",
					"per_page",
					"current_page",
					"last_page",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_Pagination_User__": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"errors": {},
					"data": {
						"$ref": "#/components/schemas/Pagination_User_"
					}
				},
				"required": [
					"code",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_User_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"files": {
						"items": {
							"$ref": "#/components/schemas/File"
						},
						"type": "array"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updateAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Enter the token with the `Bearer: ` prefix, e.g. Bearer abcde12345"
			},
			"Cookie": {
				"type": "apiKey",
				"name": "token",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "0.0.1",
		"description": "server",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_User_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Bearer": []
					},
					{
						"Cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_Pagination_User__"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Bearer": []
					},
					{
						"Cookie": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_User_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Bearer": []
					},
					{
						"Cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_User_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Bearer": []
					},
					{
						"Cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_User_"
							}
						}
					}
				}
			}
		},
		"/users/role/{id}": {
			"get": {
				"operationId": "GetUserByRoleId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_Pagination_User__"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Bearer": []
					},
					{
						"Cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}